refactor: Own module for module loader
